requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
distro==1.9.0
Django==4.0.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.5
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.1
